+++ saltus/include/saltus/vulkan/vulkan_bind_group.hh
@@ -21,8 +21,8 @@ namespace saltus::vulkan
             uint32_t binding_id,
             const std::shared_ptr<Buffer> &buffer,
             uint32_t array_index = 0,
-            uint32_t offset = 0,
-            std::optional<uint32_t> size = std::nullopt
+            uint64_t offset = 0,
+            std::optional<uint64_t> size = std::nullopt
         ) override;
 
     private:
